pyproject.toml
[project]
name = "fastapi-monitoring"
version = "0.1.0"
description = ""
authors = [{name = "K-scorpi"}]
readme = "README.md"
requires-python = ">=3.10.7,<3.14"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.7,<3.14"
fastapi = "^0.111.1"
uvicorn = "^0.29.0"
prometheus-fastapi-instrumentator = "^3.0.0"
aioredis = "^2.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
python-logstash = "^0.4.6"
requests = "^2.32.4"
pytest = "^8.4.1"
redis = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"